Extract Portfolio Assets to own component (WIP)
